tools/doc_pack/requirements.txt
markitdown>=0.1.5
openpyxl>=3.1.5
pillow>=10.4.0
PyMuPDF>=1.26.0
pymupdf4llm>=0.0.14
python-pptx>=0.6.23
